Do not attempt to infer the type for containers
`additionalProperties` and `properties` are containers that themselves
do not have a `type` property.  What use to happen is that whenever
these containers were encountered, if any of the properties within those
containers used the same name as those properties used to infer types,
the container itself would have its type inferred and that would break
the processing.

Fixes #62