Resolve issue with required field in oneOf
Having a schema that has a oneOf definition as such:

```
{
  "oneOf": [
    { required: ["foo"], /*...*/ },
    { /*...*/ },
  ]
}
```

causes the changed line to fail with error message `calling includes on undefined`. This does not always happen, as it appears that the "oneOf" field is being picked by random.

This patch simply uses a default empty array if `fixed.required` is undefined, resolving this issue.