Merge back nulls-feature into development
An example or default value set to null in a schema can make core.run function throw when traversing the schema (sub is de facto null and typeof sub.generate will throw). Adding a control on null/undefined sub fixes this. Allowing null values to be returned as a valid value (which is very common in API) has been done in the core.utils.typecast function: adding a control on null/undefined value once the value is extracted from the params makes it now possible to deal with null values (#594)