feat: enhance context with schema path (#665)
This change allows one to use the `schemaPath` from the context to get additional information from the schema. An alternative would be to copy more fields into the context, but that will eventually end up with a full schema copy in the context. By adding the `schemaPath`, it is possible for everyone to get the information they need.